implement DOS function dispatcher
it's astonishing how much abuse XACT can take. about the only thing it
doesn't like about unix-style filenames is the / filename separator, which
it dutyfully translates to \. and its FINDFIRST buffer really isn't sized
for anything beyond 8.3 filenames, though it apparently also only uses that
to check for existence of a file.

of course, that same code, minus a small C library, is also used for XACT
on ancient unix workstations. so it's somewhat expected not to assume any
DOS-isms about its filenames.