implement DOS API by binary patching the relevant libc functions
this is more portable than handling SIGSEGV. in principle it should be
possible to build xrun for Windows, though probably using Cygwin because the
DOS 2.0+ file API is closer to Unix than to Windows.

obviously binary patching is specific to the particular version of the
MetaWare High C runtime library used by Xilinx for XACT 5.20. but then
again, the only purpose of xrun is to run that toolchain, and it's not like
a toolchain that has been abandoned for 20 years now will suddenly change.

the same binary patching engine can be used to replace mwhcinit.c. that
simply hasn't been implemented yet.