handle PC speaker beep and the BIOS call trampoline
some programs emit a PC speaker beep on (usually non-critical) errors. these
shouldn't crash the entire program. that beep could be easily emulated with
a terminal beep (which, ironically, that routine is probably attempting to
emulate). but in a non-interactive pipeline, the beep is more annoying than
it is useful, so it simply gets remoted entirely.

the BIOS call trampoline is specifically used by xdelay when called without
arguments. it then tries to show its DOS-style GUI, which even in dosbox
isn't very useful. xdelay works just fine if called with arguments, but it
shouldn't just crash when run without arguments.